\name{as.matrix.partition}
\alias{as.matrix.partition}
\alias{as.partition}
\title{Coerce partitions to matrices and vice versa}
\description{
Coercion to and from partitions
}
\usage{
\S3method{as.matrix}{partition}(x, ...)
as.partition(x, ...)
}
\arguments{
  \item{x}{Object to be coerced}
  \item{\dots}{Further arguments}
}
\author{Robin K. S. Hankin}
\examples{
as.matrix(parts(5))
}
\keyword{math}
