% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gibbs.R
\name{calcError}
\alias{calcError}
\title{Calculates FNR and FDR when ground truth is available}
\usage{
calcError(z, id)
}
\arguments{
\item{z}{Matrix with posterior samples}

\item{id}{Data identifiers (ground truth)}
}
\value{
Two component vector with FNR and FDR
}
\description{
Calculates FNR and FDR when ground truth is available
}
