% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd-tree.R
\name{kdSplit}
\alias{kdSplit}
\title{A recursive nodesplitting function called in kdTree()}
\usage{
kdSplit(X, N, node, depth, max.depth)
}
\description{
A recursive nodesplitting function called in kdTree()
}

