% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nestedListToDataFrame}
\alias{nestedListToDataFrame}
\title{Convert a nested list from kdTree() to a sensible data frame}
\usage{
nestedListToDataFrame(nested.list, entry.length)
}
\arguments{
\item{nested.list}{A nested list created from kdTree()}

\item{entry.length}{The number of rows of each individual data frame}
}
\description{
Convert a nested list from kdTree() to a sensible data frame
}

