% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp-tree.R
\name{rpSplit}
\alias{rpSplit}
\title{A recursive nodesplitting function called in rpTree()}
\usage{
rpSplit(X, N, node, depth, max.depth)
}
\description{
A recursive nodesplitting function called in rpTree()
}

