% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesESC.R
\name{SampleClusterESCD}
\alias{SampleClusterESCD}
\title{Posterior samples of cluster assignments for ESCD model}
\usage{
SampleClusterESCD(data, burn, nsamples, truncds, samds = TRUE,
  spacing = 1000, w = 3, m = 10, hpriorpar = NULL, bl = NULL)
}
\arguments{
\item{data}{Data frame containing only categorical variables}

\item{burn}{MCMC burn-in period}

\item{nsamples}{MCMC iterations after burn-in}

\item{truncds}{Truncation value for distortion probabilities (betas)}

\item{samds}{Indicator to sample the fields' distortion probabilities (betas)}

\item{spacing}{Thinning for chaperones algorithm (default 1000)}

\item{w}{Step size for slice sampler (default 3)}

\item{m}{Limit on steps for slice sampler (default 10)}

\item{hpriorpar}{hyperpameter values for priors on Prior parameters}

\item{bl}{vector with block assignments for blocking method}
}
\value{
List with posterior samples for cluster assignments (Z),
        Prior parameters and distortion probabilities (Params)
}
\description{
Posterior samples of cluster assignments for ESCD model
}
