% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesESC.R
\name{SetParam}
\alias{SetParam}
\title{Parameters for reseating algortihms}
\usage{
SetParam(Prior, param, N)
}
\arguments{
\item{Prior}{Specify partition prior: "DP", "PY", "ESCNB"}

\item{param}{Vector of parameters for specified prior}

\item{N}{Number of records in the dataset}
}
\value{
Parameters for probability of existing cluster (A)
        and new cluster (B) for reseating algorithms
}
\description{
Parameters for reseating algortihms
}
\examples{
SetParam("DP", c(1), 100)
}
