\name{as.matrix.partition}
\alias{as.matrix.partition}
\title{Coerces partitions to matrices}
\description{
Coerces partitions objects to matrices
}
\usage{
as.matrix.partition(x, ...)
}
\arguments{
  \item{x}{Object to be coerced}
  \item{\dots}{Further arguments}
}
\author{Robin K. S. Hankin}
\examples{
as.matrix(parts(5))
}
\keyword{math}
